// src/api/api.ts
export interface CategoryResponseDTO {
    'id'?: string;
    'name'?: string;
    'type'?: CategoryResponseDTOTypeEnum;
    'color'?: string;
}

export const CategoryResponseDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type CategoryResponseDTOTypeEnum = typeof CategoryResponseDTOTypeEnum[keyof typeof CategoryResponseDTOTypeEnum];
// ...
export interface CreateTransactionDTO {
    'description': string;
    'date': string;
    'amount': number;
    'category_id': string;
}
export interface DashboardSummaryDTO {
    'receita'?: number;
    'despesa'?: number;
    'lucro'?: number;
    'margem'?: number;
}
// ...
export interface Pageable {
    'page'?: number;
    'size'?: number;
    'sort'?: Array<string>;
}
export interface RegisterRequestDTO {
    'username': string;
    'email': string;
    'password': string;
    'confirmPassword': string;
}
export interface ResetPasswordDTO {
    'password': string;
    'confirmPassword': string;
}
export interface TransactionImportResponseDTO {
    'validTransaction'?: number;
    'transactions'?: Array<TransactionResponseDTO>;
    'errors'?: Array<string>;
}
export interface TransactionResponseDTO {
    'id'?: string;
    'description'?: string;
    'amount'?: number;
    'date'?: string;
    'category'?: CategoryResponseDTO;
}
export interface UpdateAmountTransactionDTO {
    'amount': number;
}
export interface UpdateCategoryDTO {
    'name': string;
    'type': UpdateCategoryDTOTypeEnum;
    'color'?: string;
}

export const UpdateCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type UpdateCategoryDTOTypeEnum = typeof UpdateCategoryDTOTypeEnum[keyof typeof UpdateCategoryDTOTypeEnum];

export interface UpdateDateTransactionDTO {
    'date': string;
}
export interface UpdateDescriptionTransactionDTO {
    'description': string;
}
export interface UpdateEmailAndUsernameDTO {
    'username': string;
    'email': string;
}
export interface UpdateEmailDTO {
    'email': string;
}
export interface UpdateNameCategoryDTO {
    'name': string;
}
export interface UpdatePasswordDTO {
    'currentPassword': string;
    'password': string;
    'confirmPassword': string;
}
export interface UpdateTransactionDTO {
    'description': string;
    'date': string;
    'amount': number;
    'category_id': string;
}
// ...
export class DashboardControllerApi extends BaseAPI {
    /**
     * 
     * @param {number} year 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getCategoryTotals(year: number, options?: RawAxiosRequestConfig) {
        return DashboardControllerApiFp(this.configuration).getCategoryTotals(year, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} year 
     * @param {number} month 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getDashboardSummary(year: number, month: number, options?: RawAxiosRequestConfig) {
        return DashboardControllerApiFp(this.configuration).getDashboardSummary(year, month, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} year 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getMonthlyTotals(year: number, options?: RawAxiosRequestConfig) {
        return DashboardControllerApiFp(this.configuration).getMonthlyTotals(year, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class TransacaoApi extends BaseAPI {
    /**
     * 
     * @param {string} transactionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public _delete(transactionId: string, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration)._delete(transactionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {CreateTransactionDTO} createTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public create(createTransactionDTO: CreateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).create(createTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Pageable} pageable 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getAllPageSortingByDate(pageable: Pageable, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).getAllPageSortingByDate(pageable, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getById(transactionId: string, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).getById(transactionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {File} file 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public importXlsx(file: File, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).importXlsx(file, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateAmountTransactionDTO} updateAmountTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateAmount(transactionId: string, updateAmountTransactionDTO: UpdateAmountTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateAmount(transactionId, updateAmountTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateDateTransactionDTO} updateDateTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateDate(transactionId: string, updateDateTransactionDTO: UpdateDateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateDate(transactionId, updateDateTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateDescriptionTransactionDTO} updateDescriptionTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateDescription(transactionId: string, updateDescriptionTransactionDTO: UpdateDescriptionTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateDescription(transactionId, updateDescriptionTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} transactionId 
     * @param {UpdateTransactionDTO} updateTransactionDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateTransaction(transactionId: string, updateTransactionDTO: UpdateTransactionDTO, options?: RawAxiosRequestConfig) {
        return TransacaoApiFp(this.configuration).updateTransaction(transactionId, updateTransactionDTO, options).then((request) => request(this.axios, this.basePath));
    }
}